Add mode support to QWaylandOutput
Outputs usually have more than one mode, add an API to support them.

When a compositor creates a QWaylandOutput it has to create a
QWaylandOutputMode for each mode and set the modes list.

The current and preferred modes can be set after listing all the modes.
By default the preferred and current mode both are the highest one.

QWaylandOutputMode has read only properties so it can be used from QML
through QWaylandQuickOutput.

This commit also contains a unit test.

Change-Id: Ie3f930df6f567f31ed7325236658affa4f60e19e